refactor: extract flow reconciliation polling into useFlowReconciliation hook
- New useFlowReconciliation() hook encapsulates:
  - Exponential backoff polling (500ms → 8s)
  - Cancellation token for superseded reconciliations
  - Optimistic 'Queued' display state
  - Query cache updates (single + paginated)
  - Cleanup on unmount
- FlowCard.jsx (383 → 301 lines): removed inline sleep(), reconcileFlowAfterRun(),
  token ref management, and optimistic state — replaced with hook
- Removed unused useQueryClient and fetchFlow imports from FlowCard
- Hook is reusable for any future component that triggers flow runs
- Added cancel() method for explicit reconciliation cancellation

Resolves P1 FlowCard reconciliation item from React tech debt audit.
